backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
python-dotenv
langchain>=0.2
langchain-community>=0.2
langchain-huggingface>=0.1
langchain-ollama>=0.1.0
sentence-transformers==2.7.0
transformers>=4.44.2
accelerate
torch
faiss-cpu
pypdf
docx2txt